Add files via upload
MAN IN THE MIDDLE ATTACK(MITM) | (BASH) Raspberry pi:
•	Understanding of latest security trends, vulnerability assessment techniques and attacks like Man in the middle attack.
ARP Spoofing is the process of linking an attacker’s MAC address with the IP address of a legitimate user on a local area network using fake ARP messages. As a result, data sent by the user to the host IP address is instead transmitted to the attacker.
By using the Nmap and sniffing tool poisoned the other Ip addresses captures all the data into Wireshark networking tool. Later visualized by the third-party website to view the collected data.